[Core] Fix filtering scenarios loaded from jar
I am running cucumber tests with features loaded from classpath:, like this:

```
java -cp fatjar.jar cucumber.api.cli.Main "classpath:features/FeatureWithExamples.feature:64"
```

However, when cucumber parses this path and creates a ZipResource from it, it adds an extra slash. This becomes a problem down the line, when LinePredicate tries to filter out the pickles. Instead of filtering the pickles out, it just lets all of the though the filter.

Changed the ZipResource to preserve the original path of the zipped entity.