[Core] Handle parallel execution exceptions (#1629)
The existing parallel implementation was rather naive. It assumed that the execution of a pickle would be exception free and it assumed that it would not be interrupted. These assumptions
turn out to be unreasonable: See #1628.

By logging any exceptions thrown and re-throwing them collectively we address the first aspect and 
when interrupted all pickle executions should be interrupted as well. At this point we return control immediately.